🚧 Compile simple conjunctive operators
Compile operators that perform on values either side of them and write them into the dist buffer.

There is quite a lot of work involved in adding other types of operators that operate on values as a prefix or a post-fix (or even as a chain of parameters as in the case of Functions #1).

Finally, the joining of dist-spans also needs to be done in a way that can be reused for all of the ways described above.